+++ index.js
@@ -8,5 +8,6 @@ catchify.race = require('./race');
 catchify.reject = require('./reject');
 catchify.resolve = catchify;
 catchify.some = require('./some');
+catchify.newPromiseHandle = require('./newPromiseHandle');
 
 module.exports = catchify;